backend/requirements.txt
# Production Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
opencv-contrib-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
mediapipe==0.10.8
ultralytics==8.0.221
psutil==5.9.6
twilio==8.10.0

# Database (PostgreSQL)
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
